+++ CHANGELOG.md
@@ -38,10 +38,17 @@ and this project plans to adhere to [Semantic Versioning](http://semver.org/spec
 ### Added
 
 - The output_indexes dataset setting, which allows the user to select which features the network will be predicting (previously, it was simply 0).
+- BatchConstants Singleton class to replace global variables in batchmaker
+- Singleton class can now be deleted with the _del() method
+
+### Removed
+
+- All global variables from batchmaker
 
 ### Fixed
 
 - Type hinting in settings.py
+- Batchmaker tests no longer fail
 
 ### Added